added voting for comments, but votng for posts currently not working

+++ src/main/java/swa/labor/SocialNetwork/model/Comment.java
@@ -13,8 +13,8 @@ public class Comment {
     @Column(nullable = false)
     private String ctext;
 
-    private Float longitude;
-    private Float latitude;
+    private Double longitude;
+    private Double latitude;
 
     @Column(name = "postedAt", nullable = false)
     private LocalDateTime postedAt;
@@ -44,19 +44,19 @@ public class Comment {
         this.ctext = ctext;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
